share SKILL_IN_CONVERSATION_MODEL_ATTRIBUTES between agent message skills and conversation skills
